Fixed candle volume bug, now just to get the datetime conversion to actually work

+++ src/TDAmeritrade.hpp
@@ -8,7 +8,7 @@ namespace tda
 {
     struct Candle
     {
-        double volume;
+        int volume;
         std::pair<double, double> highLow;
         std::pair<double, double> openClose;     
         std::string datetime;
